feat(skeleton): apply theme
