+++ src/components/Skeleton/index.tsx
@@ -1,0 +1,13 @@
+import Styled from './styles';
+import { SkeletonProps } from './types';
+
+const Skeleton = ({
+    children,
+    ...rest
+}: SkeletonProps) => (
+    <Styled {...rest}>
+        {children}
+    </Styled>
+);
+
+export { Skeleton };
