+++ src/components/Skeleton/types.ts
@@ -1,0 +1,9 @@
+import { Dispatch, HTMLAttributes, SetStateAction } from 'react';
+
+export interface SkeletonProps extends HTMLAttributes<HTMLDivElement> {
+    run?: boolean;
+    onRun?: Dispatch<SetStateAction<boolean>>;
+    width?: string | number;
+    height?: string | number;
+    rounded?: boolean;
+}
