+++ src/index.ts
@@ -8,6 +8,7 @@ export * from './components/Chip';
 export * from './components/Select/Select';
 export * from './components/Card';
 export * from './components/Radio';
+export * from './components/Skeleton';
 export * from './components/Tabs/Tabs';
 export * from './components/Table/Table';
 export * from './components/Alert';
